+++ frontend/package.json
@@ -9,7 +9,6 @@
     "node-sass": "^6.0.1",
     "react": "^17.0.2",
     "react-calendar": "^3.5.0",
-    "react-datepicker": "^4.5.0",
     "react-dom": "^17.0.2",
     "react-router": "^6.0.1",
     "react-router-dom": "^6.0.1",
