+++ frontend/src/scss/popup.scss
@@ -1,0 +1,208 @@
+/*
+Colors
+*/
+$primary-1: #FFFFED;
+$primary-2: #FFFFC1;
+$primary-3-main: #FFFF81;
+$primary-4-text: #8A79AF;
+
+$secondeay-1: #ECFFFF;
+$secondeay-2-main: #DEFCFC;
+$secondeay-3-border: #CCFAFF;
+$secondeay-4: #8ED9E0;
+$secondeay-5: #49BAC6;
+$secondeay-6-text: #009EAD;
+
+$gray-white: #FFFFFF;
+$gray-1: #F5F5F5;
+$gray-2: #E6E6E6;
+$gray-3: #C2C2C2;
+$gray-4: #999999;
+$gray-5: #666666;
+$gray-6: #333333;
+$gray-black: #000000;
+
+$theme-alert: #FF8080;
+$theme-info: #FF8A00;
+
+$transparent-black: rgba(0, 0, 0, 0.05);
+
+/*
+Fonts
+*/
+@mixin headline-h5 {
+  font-family: "Inter";
+  font-size: 16px;
+  font-weight: 700;
+  line-height: 100%;
+}
+
+@mixin headline-h6 {
+  font-family: "Inter";
+  font-size: 13px;
+  font-weight: 700;
+  line-height: 100%;
+}
+
+@mixin body-p16 {
+  font-family: "Inter";
+  font-size: 16px;
+  font-weight: 500;
+  line-height: 120%;
+}
+
+@mixin body-p14 {
+  font-family: "Inter";
+  font-size: 14px;
+  font-weight: 500;
+  line-height: 120%;
+}
+
+@mixin body-p12 {
+  font-family: "Inter";
+  font-size: 12px;
+  font-weight: 500;
+  line-height: 120%;
+}
+
+@mixin button-md {
+  font-family: "Inter";
+  font-size: 16px;
+  font-weight: 600;
+  line-height: 100%;
+}
+
+@mixin button-sm {
+  font-family: "Inter";
+  font-size: 12px;
+  font-weight: 600;
+  line-height: 100%;
+}
+
+#alertPopup {
+    position: absolute;
+    display: flex; 
+    justify-content: center;
+    align-items: center;
+    top: 0px;
+    left: 0px;
+    width: 100%;
+    height: 100%;
+    background-color: $transparent-black;
+    backdrop-filter: blur(1px);
+    -webkit-backdrop-filter: blur(1px);
+    z-index: 1;
+  }
+  
+  #alertPopup.hide {
+    display: none;
+  }
+  
+  #alertPopup .content {
+    width: 420px;
+    height: 270px;
+    display: flex;
+    flex-direction: column;
+    justify-content: space-between;
+    align-items: center;
+    background-color: $primary-1;
+    border-radius: 3px;
+    box-shadow: 2px 2px 2px $transparent-black;
+    @include body-p14;
+  }
+  
+  #alertPopup .text {
+    padding-top: 70px;
+  }
+  
+  #alertPopup .btn {
+    padding-bottom: 50px;
+  }
+  
+  #confirmPopup {
+    position: absolute;
+    display: flex; 
+    justify-content: center;
+    align-items: center;
+    top: 0px;
+    left: 0px;
+    width: 100%;
+    height: 100%;
+    background-color: $transparent-black;
+    backdrop-filter: blur(1px);
+    -webkit-backdrop-filter: blur(1px);
+  }
+  
+  #confirmPopup.hide {
+    display: none;
+  }
+  
+  #confirmPopup .content {
+    width: 420px;
+    height: 270px;
+    display: flex;
+    flex-direction: column;
+    justify-content: space-between;
+    align-items: center;
+    background-color: $primary-1;
+    border-radius: 3px;
+    box-shadow: 2px 2px 2px $transparent-black;
+    @include body-p14;
+    z-index: 1;
+  }
+  
+  #confirmPopup .text {
+    padding-top: 70px;
+  }
+  
+  #confirmPopup .btn {
+    padding-bottom: 50px;
+  }
+  
+  .idxZ1 {
+    z-index: 1;
+  }
+  
+  .idxZ2 {
+    z-index: 2;
+  }
+  
+  #registerPopup {
+    position: absolute;
+    display: flex; 
+    justify-content: center;
+    align-items: center;
+    top: 0px;
+    left: 0px;
+    width: 100%;
+    height: 100%;
+    background-color: $transparent-black;
+    backdrop-filter: blur(1px);
+    -webkit-backdrop-filter: blur(1px);
+    z-index: 1;
+  }
+  
+  #registerPopup.hide {
+    display: none;
+  }
+  
+  #registerPopup .content {
+    width: 480px;
+    height: 360px;
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    background-color: $primary-1;
+    border-radius: 3px;
+    box-shadow: 2px 2px 2px $transparent-black;
+    @include body-p14;
+  }
+  
+  #registerPopup .text {
+    padding-top: 70px;
+  }
+  
+  #registerPopup .btn {
+    padding-bottom: 20px;
+  }
